refactor(storybook): move stories file outside src
